refactor(shell): extract useAppShell hook and AppLayout component
Refactored the main `App` component to improve modularity and separation of concerns, adhering to the Single Responsibility Principle.

- Extracted all state management logic (reducer, side effects, and handlers) into a new custom hook, `useAppShell`, located in `src/shell/useAppShell.ts`.
- Moved all JSX and rendering logic into a new presentational component, `AppLayout`, located in `src/shell/AppLayout.tsx`.
- Simplified the root `App` component to act as a composition layer, responsible for calling the `useAppShell` hook and rendering the `AppLayout` component within the necessary context providers.

This change enhances maintainability, readability, and testability without altering the application's functionality or appearance.
